Refactor: Created New Files And utils folder
